Refactors workflow package
